+++ baseclient.go
@@ -5,14 +5,15 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/google/go-querystring/query"
 	"io/ioutil"
 	"net/http"
 	"net/url"
 	"reflect"
+
+	"github.com/google/go-querystring/query"
 )
 
-const sdkVersion = "4.1.0"
+const sdkVersion = "4.7.1"
 
 type ClientOptions struct {
 	HttpClient *http.Client
@@ -31,15 +32,21 @@ type BaseClient struct {
 	ResponseInterceptor func(response *http.Response)
 }
 
-func (client *BaseClient) 	NewRequest(method string, params interface{}, baseRequestString string, formatString string,
-formatParams ...interface{}) (request *http.Request, err error) {
+func (client *BaseClient) NewRequest(method string, params interface{}, baseRequestString string, formatString string,
+	formatParams ...interface{}) (request *http.Request, err error) {
 
 	if client == nil || client.httpClient == nil {
 		err = errors.New("client and httpClient cannot be nil")
 		return
 	}
 
+	isCallInsightsRequest := false
+	var requestPath string
+
 	for i, param := range formatParams {
+		if !isCallInsightsRequest {
+			isCallInsightsRequest, requestPath = checkAndFetchCallInsightsRequestDetails(param)
+		}
 		if param == nil || param == "" {
 			err = errors.New(fmt.Sprintf("Request path parameter #%d is nil/empty but should not be so.", i))
 			return
@@ -47,8 +54,13 @@ formatParams ...interface{}) (request *http.Request, err error) {
 	}
 
 	requestUrl := *client.BaseUrl
-
 	requestUrl.Path = fmt.Sprintf(baseRequestString, fmt.Sprintf(formatString, formatParams...))
+
+	if isCallInsightsRequest {
+		requestUrl.Host = CallInsightsBaseURL
+		requestUrl.Path = requestPath
+	}
+
 	var buffer = new(bytes.Buffer)
 	if method == "GET" {
 		var values url.Values
@@ -58,11 +70,11 @@ formatParams ...interface{}) (request *http.Request, err error) {
 
 		requestUrl.RawQuery = values.Encode()
 	} else {
-		if(reflect.ValueOf(params).Kind().String() != "map"){
+		if reflect.ValueOf(params).Kind().String() != "map" {
 			if err = json.NewEncoder(buffer).Encode(params); err != nil {
 				return
 			}
-		}else if (reflect.ValueOf(params).Kind().String() == "map" && !reflect.ValueOf(params).IsNil()) {
+		} else if reflect.ValueOf(params).Kind().String() == "map" && !reflect.ValueOf(params).IsNil() {
 			if err = json.NewEncoder(buffer).Encode(params); err != nil {
 				return
 			}
@@ -89,7 +101,6 @@ func (client *BaseClient) ExecuteRequest(request *http.Request, body interface{}
 		return errors.New("httpClient cannot be nil")
 	}
 
-
 	response, err := client.httpClient.Do(request)
 	if err != nil {
 		return
@@ -102,9 +113,9 @@ func (client *BaseClient) ExecuteRequest(request *http.Request, body interface{}
 				err = json.Unmarshal(data, body)
 			}
 		} else {
-			if (string(data) == "{}" && response.StatusCode == 404){
+			if string(data) == "{}" && response.StatusCode == 404 {
 				err = errors.New(string("Resource not found exception \n" + response.Status))
-			}else{
+			} else {
 				err = errors.New(string(data))
 			}
 		}